Factorize 'get_public_key' code
