+++ src_features/getPublicKey/cmd_getPublicKey.c
@@ -1,7 +1,7 @@
 #include "shared_context.h"
 #include "apdu_constants.h"
 #include "common_utils.h"
-#include "feature_getPublicKey.h"
+#include "getPublicKey.h"
 #include "common_ui.h"
 #include "os_io_seproxyhal.h"
 #include "crypto_helpers.h"
