+++ src_features/getPublicKey/getPublicKey.h
@@ -3,6 +3,7 @@
 
 #include "shared_context.h"
 
+uint16_t get_public_key(uint8_t *out, uint8_t outLength);
 uint32_t set_result_get_publicKey(void);
 
 #endif  // _GET_PUB_KEY_H_
