+++ src_features/signTx/feature_signTx.h
@@ -26,7 +26,6 @@ void start_signature_flow(void);
 
 uint16_t handle_parsing_status(parserStatus_e status);
 
-uint16_t get_public_key(uint8_t *out, uint8_t outLength);
 bool max_transaction_fee_to_string(const txInt256_t *BEGasPrice,
                                    const txInt256_t *BEGasLimit,
                                    char *displayBuffer,
