+++ src_features/signTx/logic_signTx.c
@@ -8,13 +8,13 @@
 #include "common_ui.h"
 #include "ui_callbacks.h"
 #include "apdu_constants.h"
-#include "crypto_helpers.h"
 #include "format.h"
 #include "manage_asset_info.h"
 #include "handle_swap_sign_transaction.h"
 #include "os_math.h"
 #include "calldata.h"
 #include "swap_error_code_helpers.h"
+#include "getPublicKey.h"
 
 static bool g_use_standard_ui;
 
@@ -290,26 +290,6 @@ static void nonce_to_string(const txInt256_t *nonce, char *out, size_t out_size)
     uint256_t nonce_uint256;
     convertUint256BE(nonce->value, nonce->length, &nonce_uint256);
     tostring256(&nonce_uint256, 10, out, out_size);
-}
-
-uint16_t get_public_key(uint8_t *out, uint8_t outLength) {
-    uint8_t raw_pubkey[65];
-    cx_err_t error = CX_INTERNAL_ERROR;
-
-    if (outLength < ADDRESS_LENGTH) {
-        return APDU_RESPONSE_WRONG_DATA_LENGTH;
-    }
-    CX_CHECK(bip32_derive_get_pubkey_256(CX_CURVE_256K1,
-                                         tmpCtx.transactionContext.bip32.path,
-                                         tmpCtx.transactionContext.bip32.length,
-                                         raw_pubkey,
-                                         NULL,
-                                         CX_SHA512));
-
-    getEthAddressFromRawKey(raw_pubkey, out);
-    error = APDU_RESPONSE_OK;
-end:
-    return error;
 }
 
 /* Local implementation of strncasecmp, workaround of the segfaulting base implem
